requirements.txt
Flask-Admin==1.5.3
Flask-Caching==1.7.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
pbr==5.1.3
redis==3.2.1
